Minor fix
- links to reference results in tests added;
- package renamed.
