+++ .github/workflows/ci.yml
@@ -25,7 +25,6 @@ jobs:
         run: |
           python -m pip install --upgrade pip wheel setuptools
           python -m pip install -r requirements.txt
-          python -m pip list
 
       - name: Install pre-commit
         run: |
