--- src/metrics/detection/structural/__init__.py
@@ -1,4 +1,0 @@
-from src.metrics.detection.structural.average_precision import *
-from src.metrics.detection.structural.constants import *
-from src.metrics.detection.structural.orthogonal_average_precision import *
-from src.metrics.detection.structural.structural_average_precision import *
