+++ src/metrics/detection/vectorized/constants.py
