+++ src/metrics/detection/vectorized/distance/distance.py
