+++ src/metrics/detection/vectorized/distance/utils.py
